Socket code updates
Use #define's to set the fpp and fppd socket filenames.

Set the permissions on the fppd /tmp/FPPD socket so that clients
do not have to be running as root to talk to the daemon.  This is
necessary for the preceeding PHP socket code patch.

Cleanup sockets when done so that fpp running as different users
doesn't leave an unusable socket laying around.