Initial commit of Effects & Events handlers.
?Brief? explanation here for anyone wanting to play with the feature.

Effects are essentially mini-sequences.  An effect can be overlayed onto
a running sequence and will override the values for the channels the
effect is overlayed onto.  For instance, you could have a megatree in
your main sequence and play an effect at random times to make a face,
image, or words appear on the tree.

In this initial alpha version of the code, effects are stored in the
normal FPP .fseq file format but named .eseq.  This may change in the
future before merging the feature into master.  These .eseq sequence
files must have the effect starting at the beginning of the channel
space.  If your mega tree uses 3000 channels starting at channel 1500,
then the effect fseq file should have 3000 channels starting at 1.
When the effect is overlayed onto the main sequence, the starting
channel will be shifted to the proper location.  This is done to
allow effects to be used in multiple places in a display.  A
single mega tree effect could be overlayed onto multiple mega
trees of the same size.

Effects can be triggered on-demand through the web interface on the
'Events & Effects' page linked to on the menu.  They may also be
triggered via the command line using "fpp -e effectname".  Effect
sequences are stored in the sequences directory along with regular
.fseq sequence files.  They can be manipulated using the same
upload screen as normal effects.

Events have 2 components, an internal side and an external side.
The internal side of an event is used to trigger an effect to be
played on a pre-specified start channel.  Rather than having to
manually specify the start channel for the effect to display an
animated image on your mega tree, you can define an event to
display that effect on the right starting channel.  Events can
have external sides as well.  A script can be configured to run
when an event fires off.  This allows for interfacing with other
systems not supported by DMX/PixelNet.

Currently, events can be triggered the same ways as effects, via
the web interface 'Events & Effects' screen and via the command
line using "fpp -t eventname".

There are still several items on the TODO list surrounding this
feature, including the ability to trigger events via one or more
control channels in the main sequence.  Rather than having to
put the mega tree animated image effect in all your sequences,
you could just set a value in the control channel in a sequence
and the event would trigger and start the animated effect at
the desired time.