Part 1 of Sequence/Audio sync fix.
This commit was part of the effects work that had some underlying
changes to move the timer out of E131.c  The actual sync fix
will include converting this from using a timer to a thread using
nanosleep to keep in sync.  Even though this commit adds files
which are going to be removed in the next commit, it was easier
to cherry-pick this commit and apply the newer patch on top than
to have to apply the rest of this patch on top of the thread patch.

Original commit message:

Move the interval timer outside of E131.c.

The interval timer will now be set to fire whenever there is a
sequence of an effect playing.  When the last sequence or effect
ends, one more timer will be set to fire to allow blanking data
to be send to clear all channels back to 0.