Move the interval timer outside of E131.c.
The interval timer will now be set to fire whenever there is a
sequence of an effect playing.  When the last sequence or effect
ends, one more timer will be set to fire to allow blanking data
to be send to clear all channels back to 0.

Effects can now be displayed without having to have a sequence
running.