Web UI changes.
Move effects off onto their own page since there will be other
effects related items added to this page such as a list of running
effects and way to stop a running effect.

Create an 'About' page.

Change the 'Wiki/Help' menu option to 'Help' containing two sub items,
'Wiki/Help' and the About page.

Move the git version info off the top of the main page and onto the
about page.  This makes it so that we don't hit github to check the
current version of the master branch every time we load a page.  Now
we only do this when loading the About page.  This should also help
normal page loading slightly on the Pi and cut out potential lag for
hosts who can't reach the internet for one reason or another.