Fix bug with old config files exposing a crash.
Pointers for key and value weren't set to 0 because I thought they were
always being used.

When we hit the last else in processing the key, we never set value.

After the big loop checking the keys we cleanup any allocated character
arrays and since value wasn't set to null at instantiation we tried to
free it, causing a segfault.