Reintroduce SUDO as a define.
I previously removed some sudo calls that are required for things to run
properly on the PI right now.  As an easy way to toggle installations I
added SUDO as a define that can be set as "sudo" for something running
sudo (like the PI) and the empty string for a linux based install where
things are configured to run as www-data or some other user.

Eventually I'd like this to be either auto-detected or a configurable
option.