Update settings code yet again...
Another suggestion from CaptainMurdoch to initialize the settings
structure at start with default values.

Now we don't have to check for null in the getters, and will need to
free memory before configuring any of the settings.  This cleans up a
couple things in the settings code for the better.

Also, remove the mpg123 setting since we're no longer using mpg123. I
didn't add an ogg123 setting unless we end up needing it.  I'm taking
the approach that the configuring system should have the executable in
their PATH so we don't need to use absolutes in our exec calls.