Event subsystem enhancements/cleanups and command.c tweaks
Allow triggering events using an optional pair of control channels in a
sequence.  There are 2 control channels, a major and a minor.  Each
channel supports 25 values, and both channel values work together to form
an event ID (MajorID/MinorID').  This gives 25 buckets of 25 for a total
of 625 unique allowable event IDs.  Users can use the Major ID to group
events for a single or similar features into one logical group to keep
them together in the UI.  Both channels must be set to a non-zero value
to trigger an event.  Due to some sequencers munging the 0-255 data to
a 0-100 range, the control channel values are divided into ranges that
are 10 values wide.  To trigger the event with Major ID 20 and Minor ID 2,
you would set the value in the major control channel to 200 and the value
in the minor control channel to 20.

Allow event script command to contain spaces for optional arguments.

Convert command.c to allow multi-character commands.

Return an error if proper args were not supplied to a command such as the
playlist name to play, the volume level to set, etc..

Spaces to tabs in events.php for consistency within the file.