Eliminate double-GET to get FPPD status during normal operations.
In the UI, instead of calling IsFPPDrunning() then getting the status
using GetFPPstatus(), instead reverse the order.  If we can't
get the status properly, then we check to see if the daemon is
actually running.  If we did get the status successfully then we
know the daemon is running so we don't need to check.

This means we only hit the web server once every second during
normal operations instead of twice.  If fppd is not running, then
the UI will still hit the web server twice a second.  This also
eliminates having to run the 'ps' command once a second to see
if the daemon is running if it is.  This should help slightly
with page load time for the pages that display the daemon status.