Convert to using to a pthread to send light data.
Switch to using a thread to send the sequence data.

The thread is started up whenever a sequence is started.  When
the sequence finishes or is stopped, the thread will stop
automatically.

If the sequence is played along with music, the player code
will check the sync every 3 seconds.  If the sync is off,
the sleep time in the lighting thread will be tweaked slightly
to bring the sync back in line.  Once the sync comes back in
line, the sleep value will be returned back to its calculated
default.

In testing, the sync always starts off by a few frames, but
then is rather quickly pulled back inline by the code described
above.  I believe this is possibly due to audio buffering or issues
during ogg123 startup since we appear to be sending data at the
correct rate during this time.  In my testing on x86_64 hardware
so far, the test sequence sync starts out 3-4 frames out with
the audio being slightly ahead.  Within the first 10 seconds,
the sync returns to normal and stays within 1 frame of sync for
the rest of the sequence.

Committing for further testing before final cleanup of some
debug statements.  Everything appears to be working properly,
but it still needs some burn-in and testing on the Pi.