Lighting thread code tweaks after testing on the Pi.
Change the offset values used when we're out of sync from 500 and
1000 microseconds to 300 and 600.

Check the sync every second.  This happens in the main thread that
is reading in ogg123 data.  The lighting data is being processed in
its own thread and ogg123 is running in its own process, so we can
afford the performance hit in the main program thread to check the
sync once a second.  Doing this, in combination with the above
offset value changes allows my Pi to sync up the 3-4 frame difference
within 6 seconds of startup and the sequence and music stay within
+-1 frame of sync for the remaining duration of the sequence.