Update the apache config on boot up to point to the right doc root
If the user has a copy of fpp installed in /home/pi/fpp then we
need to use that instead of /opt/fpp/www so update the apache
default site config to point to the right directory.  This also
pulls in the apache config into git.

Change log directory for apache error logs to the media/logs directory
so it can be viewed through the file manager interface.