Don't log startup script output for now.
Easiest way to do this in the future would be to just tee the
output when calling the startup script from /etc/rc.local.  For
now don't redirect the whole script since we want some things
echoed to the console.  Could fix this by using a config_upgrade
script to update /etc/rc.local, but even that would have issues
if the flash drive wasn't mounted yet since we'd be logging to
the root filesystem so it's just just a simple 1-line change.