Javascript rework for some backend calls & Event editor optimization.
Hide the 'Start Channel' box on the event editor unless the user
selects an effect to run for the event.

Rework some of the Javascript calls to the webserver to wait until
the XMLHttpRequest is complete before reloading the page.  These
page reloads need to go away, but I haven't had a chance to convert
these reloads to asynchronous requests yet.

This may fix the issue some people have been seeing creating new
events.