Check to see if we should end the schedule whether repeat is set or not.
I believe this is causing several reported cases of schedules not
stopping when they should.  There is no reason I can think of to only
check to see if we should stop a playing schedule if the repeat flag is
set.  If users aren't using repeat, then we don't check to see if we
should stop and the show continues.