Don't clear the lights between sequences in a playlist and logging changes.
Previously we were sending blanking data to turn all channels off
in between each sequence in a playlist.  Now, we will only send
blanking data after the playlist is over.  This allows the user to
control which channels stay on between sequences.  In the future,
this may be a user-configurable setting in the playlist itself, but
for now we're hard-coding to not send blanking data except at the end.

This commit also has in it various logging optimizations and cleanups.

- Disable output of 's' status command query which happens once a second
  when your browser is on the status screen.
- Modify ogg123 sync code to only print sync status messages once per
  second instead of three times per second.  This will cut down on writes
  to the USB drive.
- Clean up a few other miscellaneous logging commands