Add backend support for memory mapped channel blocks.
Once the feature is complete, this will allow external scripts or
programs to overlay data onto channel output data.  Currently this
can be used to test channels from the command line as well as turn
blocks of channels on or off.

A new 'fppmm' binary has been added as a client interface to the
memory mapped channel data file.  Currently this binary can be
used to:

  - turn test mode on or off
  - set individual channel values when in test mode
  - individually enable/disable overlay of mapped blocks' data
  - copy raw data from a specified file onto the channel output data

Theoretically the 'raw data' import feature could be used to
overlay a bitmap image onto a matrix if the channels were defined
and aligned in the right direction and order.  Currently it is left
up to the external process creating this raw data file to align the
channel data up properly if the display item has folded strings, or
other configurations which cause the channel data to be out of order
or non-contiguous.