Allow users to see and download files from the 'upload' directory.
If there is a failed upload attempt, the file may get stuck
here and we also plan on using this later for other features, so
add the ability to see what files are in the 'upload' directory
and download them for other use.

The plan is also to rework the download at some point to allow
downloading any uploaded file (sequences, media, etc.).