Create a unified tabbed Channel Output config screen.
Adds support for configuring various USB dongles and
SPI-WS2801 output types.

Multiple USB dongles can now be configured and used
at the same time.

E1.31 and FPD can be disabled with a single checkbox
on their respective tabs so the users can enable/disable
without having to check/uncheck all universes.  This is
an override setting, mainly used for disabling all
interfaces in a single click.

The device list for USB dongles and SPI interfaces is
dynamically generated, so if you run this on a computer
without the SPI devices, you will not be able to add
a SPI-WS2801 output.  If there are no USB dongles
attached, then you will not be able to add any to the
config.