Makefile and instructions for building rasclock kernel module
The kernel module source is downloaded in the Makefile as part of
the build process.  The instructions were originally provided by
Materdaddy and updated by myself a little.  I have been able to
build the module using the instructions for installing a cross-
compiler noted at the top of the README.txt file.