Parse true ESEQ format *.eseq effects files.
Previously, we were parsing *.eseq effects files the same as
FSEQ format files and the user was responsible for specifying the
start channel.  Now that we have a real ESEQ file format, FPP has
been modified to parse this file format instead and the start
channel is read from the model data in the ESEQ file.  The user
can still override the start channel if desired in order to use the
same effect on multiple display features at different channel offsets.