startup script enhancements
Run /home/pi/fpp/scripts/startup if we're supposed to be running
out of /home/pi/fpp.  Previously /opt/fpp/scripts/startup was always
used.

Allow Apache to start even if the USB flash drive is not mounted.
If the logs directory does not exist, meaning that this is either
the first boot or we don't have a USB flash drive mounted, then
modify the Apache config file to put its error log in /var/tmp to
allow Apache to startup until the user inserts the flash drive or
the log directory is created later in the startup script.

Create the media/config directory on the flash drive when creating
the logs directory.  Owned by pi.pi.

Chown the media/logs directory to be owned by the pi user and group.

Change the sleep interval from 20 seconds down to 5 while waiting
for the USB flash drive to be mounted if it was not inserted when
the Pi was booted.  This should let the system boot a little
quicker (sometimes) after a flash drive is inserted.