Enable Network Configuration page in the UI.
This allows users to configure the onboard ethernet and optional
wireless NICs using either static IP information or DHCP.

If the default config on the SD card in /etc/network/interfaces is
left stock (dhcp on eth0) then the SD card should still be bootable
by removing the USB flash drive even if the network config stored
on the flash drive is messed up.  The network settings stored on
the SD card are applied first at bootup, then the config saved via
FPP is applied before FPP starts.

In order for this to work for wireless, wireless support should be
enabled in /etc/network/interfaces via the following lines:

    auto wlan0

    iface wlan0 inet dhcp
        wpa-roam /etc/wpa_supplicant/wpa_supplicant.conf

The FPP network config uses wpa_supplicant to configure the wireless
NIC.

ALL network config is stored on the flash drive in the media/config
directory, so if you take a flash drive from one Pi and put it into
another Pi, the network settings follow the flash drive.