Add an init script to initialize some things at boot time.
We want to clear out resolv.conf at boot time in case we have
a left-over static config, so this commit creates a fpp_init
script that is run by the init setup on the new SD image.

Currently the only thing the script does is clear out the
/etc/resolv.conf file, but we may find other uses for it later.
The script runs after the filesystems are mounted, but prior
to networking being started.

Fixes #32.