Allow user to enable/disable screen blanking at startup.
Add a 'Blank screen at startup' setting to the FPP settings page
to allow the user to enable or disable the screensaver setting.
The default is to not blank the screen at startup.  If users are
using projectors attached to FPP, they can enable this setting to
blank the projector screen at startup to restore the functionality
prior to this patch.