Startup cleanup and enhancements.
Wait on DHCP IP addresses for up to 60 seconds before announcing the
IP address list out the audio output.

Add a 'functions' script to contain functions we want to optionally
include in some scripts.  'common' gets included in most scripts,
so it may not be a good place for everything.

Remove execute perms from 'common' script since this is meant to be
sourced into other scripts and doesn't need to be executable.