Reset RTS when talking to DMX Open dongles.
This is mainly a mod on a patch by Stephane to clear the RTS
line to allow the Entec Open DMX dongle to work with our
USBDMXOpen.c driver.  Tested with a generic FTDI dongle and
it still works, so it seems we're safe to make this mod.

Rather than applying the mod in SerialOpen(), I am choosing
to go the safer route this close to a release and put the
code into a separate SerialResetRTS() function and call that
from USBDMXOpen_Open() rather than applying the RTS change
to all users of SerialOpen().  We may refactor again after
the release sometime.