Playlist creation fixes.
Fix adding a playlist name with a space in it.  Previously the user
would have been required to reload the page to see the correct playlist
name since fppxml.php replaces the spaces with underscores, but the UI
doesn't know about that change.  Now, we do the underscore replacement
in JavaScript so the UI is in sync.

Fix the dropdown population code when we don't have any playlists defined.
This was fixed previously but still didn't work in the first-run case where
there were no playlists defined.