Memory Mapped Channel Block enhancements.
The Memory Mapped Channel Block code now understands string
directionality and wrapping as well as how many strings make up
the block.  This allows the simplification of the logic in
client apps that would write to these blocks including the
Perl module and the UI plugin to put text on a matrix or pixel
tree both of which are near completion.

The settings are similar to Nutcracker, including the orientation,
starting corner, string count, and number of strands per string.
Eventually we plan to allow importing of this information directly
from Nutcracker.