Volume display fixes.
Move volume down to a separate line.  This gives us room to add
+/- buttons and eliminates some of the differences between the UI
for desktop and small mobile devices.  The only difference now is
that we don't display the slider on smaller screens since it is
iffy on touch devices.  Also add a numeric volume indicator so
there is visual feedback of the volume level other than the slider
which would not be visible on a mobile device.