Add a preliminary version of a documentation viewer.
but leave it commented out on the menu for now.  This supports
dropping .txt and .html files into the docs directory and they
will be listed and viewable.  Text files are wrapped in a pre
tag, html is shown as it would if viewed in a web browser
outside FPP.  Still debating how this is setup.  It can be
enabled for testing by uncommenting the Docs menu entry
in menu.inc.