Modify HexDump debug code to use standard logging.
HexDump() wasn't using our standard logging so it would not
show up in the log file.  This was mainly to allow HexDump()
to be used in binaries other than fppd.  This commit modifies
the logging code to have its own variable for the logfile name
and we set that whenever we load the settings file or parse
the command line arguments.  Now, we can use that filename to
determine whether to log to stdout or a file and the Log*
macros can be used in fpp and fppmm and any new binaries.