Change the way we set the spidev buffer size.
spidev is no longer a module in the 3.10.x raspbian kernel
packages, so we need to set the buffer size in a different
way.

This should fix FPD output on FPP v0.3.0.

Thanks to Mat for the fix, I tested on my system and it
cures the FPD output errors, but we'll need someone with a
FPD to test and confirm the FPD is working again.