[UI,Core] Add a developer.php page
This page is not available via the menus, but will allow developers or
users to switch branches.  It also shows the current 'git status' output.

When switching branches, binaries are recompiled.  Any compiled binaries
are cleaned up before switching.  Any compiled binaries which exactly
match the binary in git are removed.