[Core] Speed up startup with no DHCP on wired network
Set the eth0 interface to 'allow-hotplug' rather than 'auto' to allow
dhclient to get an IP address in the background rather than holding up
the boot process.  This speeds up boot times by around 60 seconds or
more when there is no DHCP server on the network that eth0 is connected
to or eth0 is not connected to any network.

Also sprinkle some log messages in our startup script so the user can
see what is going on during bootup.