[Events] Add UI for setting Event Control Channels
Major/Minor are the two channels for triggering events via sequence
data.  These are set to 10x the values of the Major/Minor parts of
the Event ID for the event that you want to fire off.

If you setup Channel 10 as the Major channel number and Channel
11 as the Minor channel number, then you would set these two
channels to the values 100/50 if you wanted to fire off event
10/5.  These are absolute values not percentages.  The values
10,20,30,40...200,210,220,230,240 correspond to 1-24 in the
parts of the Event ID.

Ie, if you want to fire event 10/5, you would set the