[video] Close control socket before spawning omxplayer
omxplayer spawns a dbus daemon instance which was holding port 32320
open because we weren't closing it before the omxplayer exec.  This
commit fixes this one specific case, but we really need to go through
and close all open file handles and sockets inside the forked process
so that this type of issue doesn't reoccur.