[Player] Block SIGCHLD when processing media player data
Signals are handled in the main thread.  When processing media player
data, we secure the mediaOutputLock.  If a SIGCHLD signal happens
due to media player child process finishing while we are inside the
player data parsing code, thi slock is still held.  The signal handler
is unable to secure the lock to clean up the media output handler so
a deadlock occurs.  This commit blocks SIGCHLD before we process player
data and unblocks after the lock is freed to allow the signal handler
to process any received SIGCHLD signal.

I believe this will cure the deadlock we have been experiencing with
the video output as well as the issue where the player time and status
stops updating since the main loop was deadlocked and unable to process
incoming commands from the UI or command line.