[UI] Add GPIO and Plugin log masks and make UI more understandable
UI is modified so that when you select 'all', all of the individual
checkboxes are automatically checked or unchecked.  When you select
'most', all but 'channeldata' are checked/unchecked.  If you
check an individual checkbox when 'all' or 'most' are checked,
then the 'all' or 'most' boxes are unchecked to signify that you
are no longer using the 'all' or 'most' masks.