[Playlists] Bugfixes and playlist control enhancements.
Fix a few bugs in the playlist/schedule code, some recent, some old.

Add support for controlling a running playlist via the 'fpp'
command line utility.

    To jump to the next entry in the running playlist:

    fpp -c next

    To jump to the previous entry in the running playlist:

    fpp -c prev

    To stop a running playlist:

    fpp -c stop

    To stop a running playlist gracefully:

    fpp -c graceful

These will be helpful with event scripts and allow events to
be essentially blocking by inserting a pause after the event in
the playlist and having the event script call "fpp -c next" when
the event is over to end the pause early and resume playback.