[Schedule UI] Fix ENTER when using time picker in schedule UI
The time picker set by default to auto-highlight the closest
matching time in the list.  This meant that if you hit ENTER,
the highlighted item would be selected reverting any manually typed
changes to the time value.  Turning off this auto-highlighting
allows ENTER to be used without reverting the user's manually
typed value.