[UI/Conversion] Convert files to fseq in the UI.
For this feature, it will be a 2-step process.  First, upload your file
ending in one of the xLights supported formats:
(vix|xseq|lms|las|gled|seq|hlsidata)

The file will be uploaded to the "Uploads" tab where you can highlight
the sequence and click "Convert".  You will be prompted whether you want
this file to be converted to an effect or a sequence.  As long as
everything works it should convert and flip you to the appropriate tab.
If there is a failure you should be shown any output with the word
"error" in it to give you a clue as to what went wrong.