[MultiSync] Initial MultiSync Video commit
This commit adds support for MultiSync Video.  It may need a little
tweaking to the sync code, but seems to be accurate in my testing.

This commit requires a new /usr/bin/omxplayer.bin binary which will
be posted in the forums for users wanting to test the code on a
v0.4.0 SD image.  The new binary will be included by default in v1.0.