[MemoryMap] Add support for updating text while scrolling
This commit adds support for passing a function reference to
FPP::MemoryMap->TextMessage() to allow the text message to be
updated while the text is still scrolling across the display.
This is useful for things like countdown timers, clocks, etc..

An example will be uploaded to the fpp-scripts Script Repository.