[UI] List all channel output types when adding new output.
Rather than hiding the output types that we support unless
there is appropriate hardware connected, we now will list all
types and alert the user if they try to add a channel output
for which the required hardware is not installed.  This should
be less confusing to users as they can see which outputs we
support and be informed that they need to install hardware to
use that output type.