[Core] Use hostname-specific versions of files on Remote FPP systems
On a FPP system running in Remote mode, search for a hostname-specific
version of the requested file before falling back to the exact
filename requested.  For example, when searching for a sequence named
'Sequence.fseq' on hostname 'HostName', check for the existence of
'Sequence-HostName.fseq' first and use that instead if it exists.

This will make file management easier since it allows users to
store Sequence.fseq for the master and Sequence-HostName1.fseq,
Sequence-HostName2.fseq, etc. for the Remotes.  Once we get automatic
file syncing between the Pi's being controlled Remotely, this will
be more beneficial.