[Debugging] Add a button to Pause/UnPause Sequence playback
Add a button on the developer.php page to toggle pausing and
unpausing of sequence playback.  This will be expanded in the
future on the main status page, but is added as a developer
option for now.

(cherry picked from commit bf27cdab3f99d30a1d6bbc7e773413f6cb422a63)

Conflicts:
	src/sequence.c