[Volume] Adjust omxplayer volume when amixer volume is changed
This patch attempts to keep omxplayer volume levels in sync with
music files played via mpg123/ogg123 which use the mixer settings
configured via amixer.  omxplayer requires a dB value, so one is
calculated and omxplayer is started up with that initial dB value.
When the volume level is changed in FPP while omxplayer is running,
'keystrokes' are sent to omxplayer to increase or decrease its
dB level accordingly.