[Scheduler] Fix a couple scheduler bugs.
If the current playlist is non-repeating and not playing, then
reschedule.

Adjust the end times of playlists to be non-inclusive.  This
handles the case of one playlist ending at the same time another
playlist is scheduled to start.  It does not handle overlapping
playlists if the overlap is caused by the first playlist running
over due to a sequence or media stopping gracefully.

(cherry picked from commit 324d6a5970739e8ae7567411ae0717b4641ecf83)