[LCD] Fix LCD output and add a Virtual LCD driver for debugging
LCD output was only working in standalone player mode, this fixes
it to also work in Master mode.  There was also a timeout that was
sometimes too log, so that has been bumped up to 5 seconds instead
of the default of 1 second.

A 'Virtual' LCD driver has also been added to aid in debugging.
This is only a display, the keyboard input does not yet work
fully so it is bypassed.  To test, you need to uncomment the
line which declares the VirtualLCD in fppLCD.py and comment out
the line which declares the AdaFruit LCD.