[Log ZIP] Filter out sensitive data.
I had previously been concerned with WiFi passwords being in debug log
zips that people were posting for our debugging, but those weren't
included because they are stored in a config file that isn't included in
our debug zip.

When e-mail support was added, the settings were saved in the main
settings file which is included.  I thought this would be a good time to
lok into this.  Chris and I decided a function to scrub potentially
sensitive data should be written.  Now, it's relatively easy for us to
add files to our debug zip and we can filter sensitive data out based on
the variables we pass to the scrub function.