[Networking] Single-quote SSID and PSK to allow special chars
In the interfaces.wlan0 config file, single-quote the PSK and SSID
values so that special characters like $ can be used.  Previously,
when sourcing the config file to setup the network, the $ would be
interpretted by bash as the start of a variable name and the config
generated would be broken.  Single-quoting these strings forces
bash to ignore the $ and other special characters in the PSK.