[Networking] Squash one more network config script bug
'ps -edaf' wasn't giving the full command line when run from the shell
script so it was failing to find the right PIDs for dhclient and ifplugd
sometimes leaving them running to cause havoc with systems configured
to use static IPs.  Switch to using 'ps auxw' to get the full command
line in the process list allowing us to locate the right processes.