[Events] Close control socket when running an event script
If the user runs omxplayer from an event script, this could spawn
off a dbus-daemon process which could hold the fppd control port open
preventing fppd from being restarted.  We close the socket whenever
we spawn omxplayer for video playback, so close it in the event code
as well just in case.