[Scheduler] Fix issue causing schedules to not end occasionally.
If we were too busy to call PlayListStopCheck() on the exact second
that a playlist was supposed to end, then the playlist would not be
stopped.  To fix this issue, we now check >= instead of ==.  This
situation was seen in user logs where fppd happened to be starting
the next playlist entry at the second it should be stopping and
PlayListStopCheck() was called before and after the exact second
but not on the exact second.
(cherry picked from commit aa7ae178667890c7402fe8cafd7885a0d4963096)