[plugins] Make media callback blocking.
In the "vastfmt" plugin there was a minor issue where sending the RDS
data to the device would create audio static as the song started
simultaneously.  By making the callback blocking, instead of
non-blocking the data is written out between songs and the noise is not
noticeable.

This has the side effect of "freezing" your sequence if the callback
doesn't return.  Plugins should be carefully scrutinized that they
return properly when called or a user's show could freeze.

On this note, the VastFMT plugin (which does use the media callback
type) now runs the "rds" binary using timeout, which should kill the
child process if it hasn't returned within 5 seconds.  This means even
if for some reason the rds binary hangs, your playlist should resume
within 5 seconds.