[Platform] Initial commit adding support for BeagleBone Black
This commit modifies scripts and source to check for /home/fpp and
use that directory if it exists instead of /home/pi.  The 'fpp'
user is added since FPP is no longer Pi-specific.  Also included
in this commit is the initial version of a FPP_Install.sh script
which can take a stock Debian/Raspbian system and install FPP
and it's dependencies onto the system.  This can be used for
building new FPP SD image and for setting up development systems.
Currently only Debian running on the BeagleBone Black is official
supported although the script has also received some testing on
the ODROID-C1 under Lubuntu.