[ChannelOutput] Add RPIWS281X Channel Output
This Channel Output is for driving two strings of ws281x pixels
attached to the GPIO headers on a Raspberry Pi.

There is no UI for this yet, it is being committed for further
testing while the UI code is completed.

To manually configure the Channel Output, add a config entry similar
to the following to your channeloutputs config file.

1,RPIWS281X,1,60,string1GPIO=18;string1Pixels=20

Fields are "outputEnabled,outputType,startChannel,channelCount,outputOptions"

outputOptions is:

string1GPIO (BCM GPIO Number, not pin number)
string1Pixels (# of pixels attached on string #1)
string2GPIO (BCM GPIO #)
string2Pixels (# of pixesl attached on string #2)